move create_dataset.py to dataset/
